backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
web3==5.24.0
flake8==7.0.0
python-dotenv==1.0.1